refactor: add invariant helpers
